# benchmark: add auto-compact flags