For https://github.com/coreos/etcd/issues/4633.