@xiang90 2016-03-01
can you please try to get some initial benchmark result out of this? So we know this is useful?