@gyuho 2016-03-03
@xiang90 PTAL. Added `same-key` option:

Results:

```
# without compaction
benchmark --conns=10 --clients=1000 put --total=1000000 --same-key

Summary:
  Total:        96.4195 secs.
  Slowest:      0.2931 secs.
  Fastest:      0.0179 secs.
  Average:      0.0964 secs.
  Stddev:       0.0265 secs.
  Requests/sec: 10371.3499

Response time histogram:
  0.018 [1]     |
  0.045 [5966]  |
  0.073 [183154]        |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.100 [416139]        |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.128 [279370]        |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.156 [92690] |∎∎∎∎∎∎∎∎
  0.183 [17999] |∎
  0.211 [2784]  |
  0.238 [915]   |
  0.266 [550]   |
  0.293 [432]   |

Latency distribution:
  10% in 0.0650 secs.
  25% in 0.0777 secs.
  50% in 0.0935 secs.
  75% in 0.1120 secs.
  90% in 0.1308 secs.
  95% in 0.1431 secs.
  99% in 0.1687 secs.
```

```
# with compaction
benchmark --conns=10 --clients=1000 put --total=1000000 --same-key --compact-index-delta=1000 --compact-interval=0.5s

Summary:
  Total:        102.2650 secs.
  Slowest:      0.5117 secs.
  Fastest:      0.0194 secs.
  Average:      0.1022 secs.
  Stddev:       0.0283 secs.
  Requests/sec: 9779.9628

Response time histogram:
  0.019 [1]     |
  0.069 [76991] |∎∎∎∎
  0.118 [681804]        |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.167 [224576]        |∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.216 [15464] |
  0.266 [310]   |
  0.315 [2]     |
  0.364 [0]     |
  0.413 [233]   |
  0.462 [206]   |
  0.512 [561]   |

Latency distribution:
  10% in 0.0715 secs.
  25% in 0.0841 secs.
  50% in 0.0990 secs.
  75% in 0.1170 secs.
  90% in 0.1362 secs.
  95% in 0.1489 secs.
  99% in 0.1760 secs.
```

So in this specific case, compaction has `-5.7%` of impact in terms of `request/sec`.