@xiang90 2016-03-03
5% seems to be a pretty small number. Have you tried this multiple times? Is the result stable? The overhead is always 5%?