@xiang90 2016-03-03
also in the test, we can put 10k keys per second. but we only compacts 1000 revision per second. the compaction will not be able to keep up with the key put.