@gyuho 2016-03-03
@xiang90 I tried several times with 100k keys and they all vary between 3 ~ 7% (slower in auto-compaction side).