@xiang90 2016-03-03
oh. i was wrong. it is the delta not the compaction increase.