@gyuho 2016-03-03
Yeah compaction is 10k as fixed, I think. 
Btw. just deleted `same-key` flag.