@xiang90 2016-03-03
/cc @mitake Can you take a look at this? maybe play with this a little bit?