@mitake 2016-03-04
@gyuho @xiang90 I tried to execute the new flags related to compaction but I got the below error:
```
$ ./benchmark --conns=10 --clients=1000 put --total=1000000 --compact-index-delta=1000 --compact-interval=0.5s
6200 / 1000000 Bom                                                                                                                                                                                ! 0.62 % 2m8spanic: rpc error: code = 11 desc = "etcdserver: storage: required revision has been compacted"

goroutine 1141 [running]:
panic(0xa71180, 0xc82088a6c0)
        /usr/local/go/src/runtime/panic.go:464 +0x3e6
github.com/coreos/etcd/tools/benchmark/cmd.compactKV(0xc8202ec000, 0x3e8, 0x3e8)
        /home/mitake/gopath/src/github.com/coreos/etcd/tools/benchmark/cmd/put.go:150 +0x389
github.com/coreos/etcd/tools/benchmark/cmd.putFunc.func2(0xc8202ec000, 0x3e8, 0x3e8)
        /home/mitake/gopath/src/github.com/coreos/etcd/tools/benchmark/cmd/put.go:103 +0x45
created by github.com/coreos/etcd/tools/benchmark/cmd.putFunc
        /home/mitake/gopath/src/github.com/coreos/etcd/tools/benchmark/cmd/put.go:105 +0x526
```

I tried `compact` branch of @gyuho 's repository. Could you take a look, @gyuho ?