@mitake 2016-03-04
@gyuho BTW I'm using V3DemoProcfile on single machine (of course I should test on a cluster later). Should I change a way of booting etcd servers?