@gyuho 2016-03-04
@mitake What was the error message that you got from `Compact`?