@gyuho 2016-03-04
Oh I see

```
"etcdserver: storage: required revision has been compacted"
```