@gyuho 2016-03-04
@mitake  So this happens when:

1. compact at revision 100
2. put... current revision 1090
3. compact at revision 90 (1090-1000)
4. 90 < 100, so returns error

It means during the compact interval, it didn't proceed fast enough to cover the compact delta.
I will cache the previous compact revision and compare two revs to prevent this.