@gyuho 2016-03-04
@mitake Currently there's no way to prevent it other than just print out the logs.
Can you try this again? And I recommend to pass smaller `compact-index-delta`
if you keep seeing errors. (I changed it from panic to error with log.Println)

Please let me know how it goes.

Thanks.