@mitake 2016-03-04
@gyuho I still got errors like below even with `compact-interval-delta == 1`:

```
602 / 1000000 Bm                                                                                                                                                                                 ! 0.06 % 22m9s2016-03-04 16:22:03.875702 I | error (rpc error: code = 11 desc = "etcdserver: storage: required revision has been compacted") when current revision 112404, revision to compact 112304
```

How about simply breaking the compaction request loop after success? Current loop issues compaction requests to single revision so it is causing the errors. This change solved the error: https://github.com/mitake/etcd/commit/55c27a42e0feb9dfcf29b17b3ec789cacb0f83c7