@mitake 2016-03-04
BTW, should compaction request to already compacted revisions be treated as an error? I'm still not sure but how about finish without error such a request simply?