@gyuho 2016-03-04
@mitake You are right. Compact goes through raft so we need to break after success. Just pushed fix. Can you try again? You can ignore the logs if you get only a few. Try to minimize the errors, so that we can maximize the compaction runs during benchmark.

Thanks.